Create struct to allow reuse code for testing
